Disable nat outside on gw interface for VPNaaS
For VPNaaS routers we don't want to have "ip nat outside" on the gateway
interface, as this will cause outgoing connections to get natted, which
will break bidirectional connectivity. As we don't need NAT on our
VPNaaS routers anyway we just remove the statement.

Detection of "this is a VPNaaS" router is done in the agent by checking
if a) we have VPNs on the router and b) that there are no internal
subnets attached. With the current feature set we don't allow internal
subnets on an VPNaaS router, but just in case I would rather want to
break the VPN instead of having unnatted traffic from an internal
subnet. If a router accidentally gets a VPN that shouldn't have one then
I'd opt for the internal subnet instead of the VPN (and this would need
fixing anyway).

This change might result in clients being able to spoof traffic by
sending traffic to the router via the VPN tunnel that then egresses the
router via the default route with a source ip controlled by the user. We
will take care of this in a followup commit.