Spoofing protection for VPNaaS router
If we are in VPNaaS mode (VPNs present + no internal subnets) we disable
NAT on the gateway interface (no ip nat outside). This could potentially
lead to spoofing: If a user sends packets through the tunnel, which then
are not routed to the BGPVPN this traffic might get routed via default
route to the external network of the gateway, thus resulting in the
fabric (in our case mostly ACI) learning the source ip of the packet
controlled by the user as an endpoint on this router. To prevent this we
want to only allow traffic to egress the router that originates from the
router itself. This is achieved by adding an ACL to the router that
contains only the router's own IPs.

This also avoids leaking internal tunnel traffic to the outside. Another
option to solve only the tunnel leaking would be to mark incoming
traffic from Tunnel interfaces and let the gateway interface drop that
packet - much like we do with our gateway interface looping protection.